Add files via upload
A deepfake is a synthetic media in which a person in an existing image or video is replaced with someone else's likeness. This can be done using a technique called deep learning, which allows computers to learn to mimic the appearance and behavior of real people. Deepfakes can be used for a variety of purposes, including entertainment, satire, and disinformation.

Here are some examples of deepfakes:

A video of President Obama saying that he supports a particular candidate in the upcoming election.
A photo of a celebrity endorsing a product.
A video of a person saying or doing something that they never said or did.
Deepfakes can be difficult to detect, as they can be very realistic. However, there are a number of techniques that can be used to identify deepfakes